Refactor utils package
