+++ iamctl/cmd/interactive/Util.go
@@ -28,7 +28,7 @@ import (
 	"os"
 	"strings"
 
-	"github.com/wso2-extensions/identity-tools-cli/iamctl/pkg/utils/interactive"
+	"github.com/wso2-extensions/identity-tools-cli/iamctl/pkg/utils"
 )
 
 type ServerInfo struct {
