+++ iamctl/cmd/interactive/add.go
@@ -24,7 +24,7 @@ import (
 	"net/url"
 
 	"github.com/spf13/cobra"
-	"github.com/wso2-extensions/identity-tools-cli/iamctl/pkg/utils/interactive"
+	"github.com/wso2-extensions/identity-tools-cli/iamctl/pkg/utils"
 )
 
 var createUsingCommand = &cobra.Command{
