+++ iamctl/cmd/interactive/createBasicApp.go
@@ -26,7 +26,7 @@ import (
 	"log"
 	"net/http"
 
-	"github.com/wso2-extensions/identity-tools-cli/iamctl/pkg/utils/interactive"
+	"github.com/wso2-extensions/identity-tools-cli/iamctl/pkg/utils"
 )
 
 type ServiceProvider struct {
