+++ iamctl/cmd/interactive/init.go
@@ -29,7 +29,7 @@ import (
 	"github.com/mbndr/figlet4go"
 	"github.com/spf13/cobra"
 	"github.com/wso2-extensions/identity-tools-cli/iamctl/cmd"
-	"github.com/wso2-extensions/identity-tools-cli/iamctl/pkg/utils/interactive"
+	"github.com/wso2-extensions/identity-tools-cli/iamctl/pkg/utils"
 )
 
 var sampleSPCmd = &cobra.Command{
